requirements.txt
torch==2.3.1+cu121
tqdm==4.66.4
safetensors==0.4.3
pillow==10.2.0
